add 1 test to protocol_test for liquid_handle_dispense and move the rest of lhd tests to separate test file
