+++ test/protocol_test.py
@@ -27,6 +27,7 @@ from autoprotocol.instruction import (
 )
 from autoprotocol.protocol import Protocol, Ref
 from autoprotocol.unit import Unit, UnitError
+from autoprotocol.liquid_handle.dispense import Dispense as DispenseMethod
 
 
 class TestProtocolMultipleExist(object):
@@ -3624,3 +3625,24 @@ class TestEvaporate(object):
             },
         )
         assert self.p.instructions[0].op == "evaporate"
+
+
+class LiquidHandleTester(object):
+    @pytest.fixture(autouse=True)
+    def setup(self):
+        self.protocol = Protocol()
+        self.tube = self.protocol.ref("tube", cont_type="micro-2.0", discard=True)
+        self.tube.well(0).set_volume("50:microliter")
+        self.flat = self.protocol.ref("flat", cont_type="96-flat", discard=True)
+
+class TestLiquidHandleDispenseMode(LiquidHandleTester):
+    def test_location_count(self):
+        volume = "5:uL"
+        instruction = self.protocol.liquid_handle_dispense(
+            source=self.tube.well(0),
+            destination=self.flat.wells_from(0, 12),
+            volume=volume,
+        )
+        assert len(instruction.data["locations"]) == 15
+        assert self.protocol.instructions[-1].op == "liquid_handle"
+
